refactor: :truck: move stories to own folder
