+++ src/stories/Layout.stories.tsx
@@ -1,21 +1,23 @@
 // boilerplate story example
 
 import { ComponentMeta, ComponentStory } from '@storybook/react'
-import Header, { IHeader } from '.'
+import Layout from '../components/Layout'
+import { ILayout } from '../interfaces/Layout.interface'
 
 export default {
-  title: 'Header',
-  component: Header,
+  title: 'Layout',
+  component: Layout,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof Header>
+} as ComponentMeta<typeof Layout>
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Header> = (args) => <Header {...args} />
+const Template: ComponentStory<typeof Layout> = (args) => <Layout {...args} />
 
 export const Base = Template.bind({})
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  label: 'Header',
-} as IHeader
+  label: 'Layout',
+  path: 'home',
+} as ILayout
